requirements.txt
# This file may be used to create an environment using:
# $ pip install -r requirements.txt
# platform: unix
absl-py==1.4.0
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna==3.4
markdown==3.4.3
numpy==1.24.3
oauthlib==3.2.2
pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
torch==2.0.1
pyyaml==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.13.0
torchvision==0.15.2
werkzeug==2.3.4
absl-py==1.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.25.2
hyperframe==6.0.1
idna==3.4
info-nce-pytorch==0.1.4
Jinja2==3.1.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
modal==0.64.199
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numpy==1.24.3
oauthlib==3.2.2
packaging==24.1
Pillow==9.5.0
propcache==0.2.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.2
rsa==4.9
safetensors==0.4.5
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
sympy==1.13.3
synchronicity==0.8.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
timm==1.0.9
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.5
typer==0.12.5
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
urllib3==1.26.3
watchfiles==0.24.0
Werkzeug==2.3.4
yarl==1.15.5
